# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Firestick API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/indexing/run": {
      "get": {
        "summary": "Trigger indexing run",
        "parameters": [
          { "name": "root", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "excludeDirs", "in": "query", "schema": { "type": "string" } },
          { "name": "excludeGlobs", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Indexing started", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IndexingReport" } } } },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "summary": "Trigger indexing run (JSON)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IndexingRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Indexing started", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IndexingReport" } } } },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/indexing/cancel": {
      "post": {
        "summary": "Request cancellation of an indexing job",
        "parameters": [
          { "name": "jobId", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1 }, "description": "If omitted, cancels the latest RUNNING job" }
        ],
        "responses": {
          "202": { "description": "Cancellation requested" },
          "404": { "description": "Job not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "No running job to cancel", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "IndexingRequest": {
        "type": "object",
        "required": ["rootPath"],
        "properties": {
          "rootPath": { "type": "string", "minLength": 1 },
          "excludeDirNames": { "type": "array", "items": { "type": "string" } },
          "excludeGlobPatterns": { "type": "array", "items": { "type": "string" } }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "path": { "type": "string" },
          "requestId": { "type": "string", "description": "Correlation id; also returned in X-Request-Id header" },
          "status": { "type": "integer" },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
      "IndexingReport": {
        "type": "object",
        "properties": {
          "jobId": { "type": "integer" },
          "status": { "type": "string" }
        }
      },
...
    }
  }
}